smtpd/requirements.txt
aiosmtpd
python-dotenv
pymongo